requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
aiofiles>=23.2.1
PyYAML>=6.0.1

# Raspberry Pi GPIO (zainstaluj tylko na Raspberry Pi)
# RPi.GPIO>=0.7.1
# gpiozero>=2.0

# GUI
# tkinter jest wbudowany w Python
# Dla wykresów (opcjonalnie):
# matplotlib>=3.8.0

# Development
python-multipart>=0.0.6
